Create synthetic_data_generation
This code was created to generate synthetic data to demo 211 call center forecast app. It was designed to run on Google Colab so that everyone can run it but can easily be converted to be run elsewhere. Please make sure to modify the path for the output file and ensure you have faker installed in your environment.